+++ CHANGELOG.md
@@ -266,6 +266,11 @@ Better utilities (for Promptbase app)
 
 Add new OpenaAI models `gpt-4o` and `gpt-4o-2024-05-13`
 
+-   Add model `gpt-4o`
+-   Add model `gpt-4o-2024-05-13`
+-   Classes that implements `LlmExecutionTools` must expose compatible models <!-- [♐] -->
+-   List OpenAI models dynamically
+
 ## In prepare
 
 ### `0..0` _(2024-0-)_